+++ frontend/src/lib/components/ScheduleEditor.svelte
@@ -69,6 +69,12 @@
 
 	$: script_path != '' && loadScript(script_path)
 
+	// set isValid to true when a script/flow without any properties is selected
+	$: runnable?.schema &&
+		runnable.schema.properties &&
+		Object.keys(runnable.schema.properties).length === 0 &&
+		(isValid = true)
+
 	const dispatch = createEventDispatcher()
 
 	async function loadScript(p: string | undefined): Promise<void> {
@@ -245,8 +251,6 @@
 				/>
 			{/if}
 			<div class="mt-6">
-				<span class="font-semibold text-gray-700">Arguments</span>
-
 				{#if runnable}
 					{#if runnable?.schema && runnable.schema.properties && Object.keys(runnable.schema.properties).length > 0}
 						<SchemaForm disabled={!can_write} schema={runnable.schema} bind:isValid bind:args />
@@ -264,7 +268,7 @@
 			<h2 class="border-b pb-1 mt-8 mb-2">Error Handler</h2>
 
 			<ScriptPicker
-				disabled={initialScriptPath != '' || !can_write}
+				disabled={!can_write}
 				initialPath={errorHandlerPath}
 				kind={Script.kind.SCRIPT}
 				allowFlow={true}